+++ xtask/src/generate/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod tests;
